requirements.txt
# Confluence + data processing
requests==2.31.0
python-dotenv==1.0.1
redis==5.0.1
openai==1.40.6
sentence-transformers==2.2.2
nltk==3.8.1
ollama==0.3.1
beautifulsoup4==4.12.3
huggingface-hub==0.23.5
tiktoken==0.7.0
numpy==1.26.4
fastapi==0.115.0
uvicorn[standard]==0.30.6

### Not used
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-ollama==0.2.0

